+++ spec/integration/fixtures/rubocop/right_hand.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+def right_hand(a, b)
+  a.divmod(b) => x, y => k, l
+  k + x => z
+
+  y + l + z
+end